feat(ux): improve accessibility of currency exchange filters
-   Updated `ParseableInputBox` to support `id`, `placeholder`, `aria_label`, and `aria-invalid` props.
-   Refactored `FilterCard` and `FilterModal` in `currency_exchange.rs` to use `<label>` elements for input fields, providing implicit labeling.
-   Added placeholders to currency exchange filter inputs for better usability.
-   Fixed a compilation error in `related_items.rs` involving a `u16` to `u32` type mismatch.

+++ ultros-frontend/ultros-app/src/components/related_items.rs
@@ -306,7 +306,7 @@ fn Recipe(recipe: &'static Recipe, item_id: ItemId) -> impl IntoView {
 
 fn npc_rows(npc: &ENpcBase) -> impl Iterator<Item = u32> + '_ {
     // TODO- can I just parse the csv into a vec?
-    npc.e_npc_data.iter().map(|i| i.0)
+    npc.e_npc_data.iter().map(|i| u32::from(i.0))
 }
 
 fn gil_shop_to_npc(gil_shops: &[GilShopId]) -> Vec<(GilShopId, &'static ENpcBase)> {
